MB-10533 Move env-related class + funcs to utils/base.py
They don't really belong in hosts.py. Not sure if maybe they should be on their own.
